+++ frontend/package.json
@@ -34,8 +34,11 @@
     "react-singleton-hook": "^4.0.1",
     "zustand": "^4.5.2"
   },
+  "overrides": {
+    "near-api-js": "^3.0.4"
+  },
   "devDependencies": {
-    "eslint": "^8.41.0",
-    "eslint-config-next": "14.2.0"
+    "eslint": "^8.57.0",
+    "eslint-config-next": "14.2.2"
   }
 }
