+++ frontend/src/config.js
@@ -1,5 +1,4 @@
 const contractPerNetwork = {
-  mainnet: 'hello.near-examples.near',
   testnet: 'coinflip.near-examples.testnet',
 };
