--- frontend_old/assets/global.css
@@ -1,167 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-html {
-  --bg: #efefef;
-  --fg: #1e1e1e;
-  --gray: #555;
-  --light-gray: #ccc;
-  --shadow: #e6e6e6;
-  --success: rgb(90, 206, 132);
-  --primary: #FF585D;
-  --secondary: #0072CE;
-
-  background-color: var(--bg);
-  color: var(--fg);
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Helvetica, Arial, sans-serif;
-  font-size: calc(0.9em + 0.5vw);
-  line-height: 1.3;
-}
-
-body {
-  margin: 0;
-  padding: 1em;
-}
-
-main {
-  margin: 0 auto;
-  max-width: 26em;
-}
-
-h1 {
-  background-image: url(./logo-black.svg);
-  background-position: center 1em;
-  background-repeat: no-repeat;
-  background-size: auto 1.5em;
-  margin-top: 0;
-  padding: 3.5em 0 0;
-  text-align: center;
-  font-size: 1.5em;
-}
-
-h2 {
-  text-align: center;
-}
-
-table {
-  margin-top: 1em;
-}
-
-.status {
-  font-size: small;
-  font-family: monospace;
-  text-align: center;
-}
-
-.please-wait .change {
-  pointer-events: none;
-}
-
-
-a,
-.link {
-  color: var(--primary);
-  text-decoration: none;
-}
-
-a:hover,
-a:focus,
-.link:hover,
-.link:focus {
-  text-decoration: underline;
-}
-
-a:active,
-.link:active {
-  color: var(--secondary);
-}
-
-button,
-input {
-  font: inherit;
-  outline: none;
-}
-
-button {
-  background-color: var(--secondary);
-  border-radius: 5px;
-  border: none;
-  color: #efefef;
-  cursor: pointer;
-  padding: 0.3em 0.75em;
-  transition: transform 30ms;
-}
-
-button:focus {
-  border: 2px solid rgb(48, 188, 244);
-}
-
-button:hover {
-  box-shadow: 0 0 10em rgba(255, 255, 255, 0.2) inset;
-}
-
-@media (prefers-color-scheme: dark) {
-  html {
-    --bg: #1e1e1e;
-    --fg: #efefef;
-    --gray: #aaa;
-    --shadow: #2a2a2a;
-    --light-gray: #444;
-  }
-
-  h1 {
-    background-image: url(./logo-white.svg);
-  }
-
-  input:focus {
-    box-shadow: 0 0 10em rgba(255, 255, 255, 0.02) inset;
-  }
-}
-
-/* coin flip animations: https://rezabaharvand.dev/blog/coin-flip-javascript */
-
-#coin {
-  height: 150px;
-  width: 150px;
-  margin: 2em auto 0 auto;
-  transform-style: preserve-3d;
-}
-
-#coin img {
-  width: 150px;
-}
-
-.heads {
-  transform: rotateY(180deg);
-}
-
-.tails,
-.heads {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  backface-visibility: hidden;
-}
-
-@keyframes flip-heads {
-  100% {
-    transform: rotateY(540deg);
-  }
-}
-
-@keyframes flip-tails {
-  100% {
-    transform: rotateY(360deg);
-  }
-}
-
-@keyframes flip {
-  0% {
-    transform: rotateY(0);
-  }
-
-  100% {
-    transform: rotateY(2160deg);
-  }
-}
