--- frontend_old/index.html
@@ -1,58 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="icon" href="assets/favicon.ico"/>
-  <link rel="apple-touch-icon" href="assets/favicon.ico"/>
-  <!-- <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous"> -->
-  <title>Welcome to NEAR</title>
-  <link rel="stylesheet" href="assets/global.css">
-</head>
-<body>
-  <div id="signed-in-flow">
-    <button style="float: right" id="sign-out-button">
-      Sign out <span data-behavior="account-id"></span>
-    </button>
-  </div>
-  <main>
-    <h1 id="signed-out-flow"> Welcome! Login to Play </h1>
-    <h1 id="signed-in-flow"> Welcome <span data-behavior="account-id"></span>!</h1>      
-    
-    <div id="coin">
-      <div class="tails">
-        <img src="./assets/tails.png" alt="Coin's tail" />
-      </div>
-      <div class="heads">
-        <img src="./assets/heads.png" alt="Coin's head" />
-      </div>
-    </div>
-
-    <div id="signed-out-flow">
-
-      <p style="text-align: center; margin-top: 2.5em">
-        <button id="sign-in-button">Sign in with NEAR Wallet</button>
-      </p>
-
-    </div>
-
-    <div id="signed-in-flow">
-      
-      <table>
-        <tr>
-          <td><p>What do you think is coming next?</p></td>
-          <td>
-              <button id="choose-tails">Tails</button>
-              <button id="choose-heads">Heads</button>
-          </td>
-        </tr>
-      </table>
-      <p class="status"><b>Status</b>: <span data-behavior="status">Waiting user input</span></p>
-      <h3>
-        Your points so far: <span class="points" data-behavior="points"></span>
-      </h3>
-    </div>
-  </main>
-  <script src="index.js" type="module"></script>
-</body>
-</html>
